+++ packages/log/src/index.ts
@@ -3,6 +3,10 @@
  */
 import { logged } from './utils';
 
+import type { MarkOptions, MeasureOptions } from './types';
+
+export type { MeasureOptions };
+
 /**
  * Logs a message with `message` if environment is not `production`.
  *
@@ -127,87 +131,6 @@ export function createTiming(
 	};
 }
 
-interface UploadTimingOptions {
-	measureName: string;
-	startTime: number | string;
-	endTime?: number | string;
-	// The color the entry will be displayed with in the timeline. Can only be a value from the
-	// palette defined in DevToolsColor
-	color?: DevToolsColor;
-	// The name (and identifier) of the extension track the entry belongs to. Entries intended to
-	// be displayed to the same track should contain the same value in this property.
-	track?: string;
-	// A short description shown over the entry when hovered.
-	hintText?: string;
-	// key-value pairs added to the details drawer when the entry is selected.
-	detailsPairs?: [ string, string ][];
-}
-
-// export function getUploadTiming(
-// 	{
-// 		measureName,
-// 		startTime,
-// 		endTime = performance.now(),
-// 		hintText,
-// 		detailsPairs = [],
-// 		color = 'primary',
-// 		track = 'Media Experiments',
-// 	}: UploadTimingOptions
-// ) {
-// 	const measure = {
-// 		start: startTime,
-// 		end: performance.now(),
-// 		detail: {
-// 			devtools: {
-// 				metadata: {
-// 					extensionName: "React Extension",
-// 					dataType: "track-entry",
-// 				},
-// 				color: "tertiary-light",
-// 				track: "An Extension Track",
-// 				hintText: "This is a rendering task",
-// 				detailsPairs: [
-// 					["Description", "This is a child task"],
-// 					["Tip", "Do something about it"],
-// 				],
-// 			},
-// 		},
-// 	};
-//
-// 	return {
-// 		name: measureName,
-// 		measure: measureOptions,
-// 	};
-// }
-
-type DevToolsColor =
-	| 'primary'
-	| 'primary-light'
-	| 'primary-dark'
-	| 'secondary'
-	| 'secondary-light'
-	| 'secondary-dark'
-	| 'tertiary'
-	| 'tertiary-light'
-	| 'tertiary-dark'
-	| 'error';
-
-interface MeasureOptions {
-	measureName: string;
-	startTime: number | string;
-	endTime?: number | string;
-	// The color the entry will be displayed with in the timeline. Can only be a value from the
-	// palette defined in DevToolsColor
-	color?: DevToolsColor;
-	// The name (and identifier) of the extension track the entry belongs to. Entries intended to
-	// be displayed to the same track should contain the same value in this property.
-	track?: string;
-	// A short description shown over the entry when hovered.
-	hintText?: string;
-	// key-value pairs added to the details drawer when the entry is selected.
-	detailsPairs?: [ string, string ][];
-}
-
 export function measure( {
 	measureName,
 	startTime,
@@ -217,6 +140,10 @@ export function measure( {
 	color = 'primary',
 	track = 'Media Experiments',
 }: MeasureOptions ) {
+	if ( ! isDev() ) {
+		return;
+	}
+
 	performance.measure( measureName, {
 		start: startTime,
 		end: endTime,
@@ -240,20 +167,6 @@ export function measure( {
 	} );
 }
 
-interface MarkOptions {
-	markName: string;
-	// The color the entry will be displayed with in the timeline. Can only be a value from the
-	// palette defined in DevToolsColor
-	color?: DevToolsColor;
-	// The name (and identifier) of the extension track the entry belongs to. Entries intended to
-	// be displayed to the same track should contain the same value in this property.
-	track?: string;
-	// A short description shown over the entry when hovered.
-	hintText?: string;
-	// key-value pairs added to the details drawer when the entry is selected.
-	detailsPairs?: [ string, string ][];
-}
-
 export function mark( {
 	markName,
 	hintText,
@@ -261,6 +174,10 @@ export function mark( {
 	color = 'primary',
 	track = 'Media Experiments',
 }: MarkOptions ) {
+	if ( ! isDev() ) {
+		return;
+	}
+
 	performance.mark( markName, {
 		detail: {
 			devtools: {
